Split __init__ out into separate submodules
