+++ src/ducktools/classbuilder/__init__.pyi
@@ -1,19 +1,44 @@
+# MIT License
+#
+# Copyright (c) 2024-2026 David C Ellis
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+
 import sys
-import types
 import typing
-import typing_extensions
 
 __lazy_modules__: list[str]
 
-from collections.abc import Callable, Iterable, Mapping
+from collections.abc import Callable, Mapping
 from types import MappingProxyType
+
+from .constants import _NothingType
+from .methods import MethodMaker, _SignatureMaker
 
 if sys.version_info >= (3, 14):
     import reannotate
+    import annotationlib
 
-    _py_type = reannotate.DeferredAnnotation | type | str
+    _private_type = reannotate.DeferredAnnotation | type | str
+    _field_type = annotationlib.ForwardRef | type | str
 else:
-    _py_type = type | str
+    _private_type = _field_type = type | str
 
 _CopiableMappings = dict[str, typing.Any] | MappingProxyType[str, typing.Any]
 
@@ -24,74 +49,11 @@ _gatherer_returntype = tuple[dict[str, Field], dict[str, typing.Any]]
 
 __version__: str
 __version_tuple__: tuple[str | int, ...]
-INTERNALS_DICT: str
-META_GATHERER_NAME: str
-GATHERED_DATA: str
-REPLACE_NAME: str
+
 
 @typing.type_check_only
 class GetFieldsProtocol(typing.Protocol):
     def __call__(self, cls: type, *, local: bool = ...) -> Mapping[str, Field]: ...
-
-def get_fields(cls: type, *, local: bool = ...) -> dict[str, Field]: ...
-def get_flags(cls: type) -> dict[str, bool]: ...
-def get_methods(cls: type) -> types.MappingProxyType[str, MethodMaker]: ...
-def get_generated_code(cls: type) -> dict[str, GeneratedCode]: ...
-def print_generated_code(cls: type) -> None: ...
-def build_completed(cls: type) -> bool: ...
-
-class _NothingType:
-    custom: str | None
-    def __new__(cls, custom: str | None = ...) -> typing.Self: ...
-    def __repr__(self) -> str: ...
-
-NOTHING: _NothingType
-FIELD_NOTHING: _NothingType
-
-class _KW_ONLY_META(type):
-    def __repr__(self) -> str: ...
-
-class KW_ONLY(metaclass=_KW_ONLY_META): ...
-
-# Stub Only Protocols
-@typing.type_check_only
-class _CodegenType(typing.Protocol):
-    def __call__(self, cls: type, funcname: str = ...) -> GeneratedCode: ...
-
-@typing.type_check_only
-class _InitArgcountCodegenType(typing.Protocol):
-    def __call__(
-        self,
-        argcount: int,
-        frozen: bool,
-        frozen_and_slotted: bool,
-        /,
-        *,
-        funcname: str = ...
-) -> GeneratedCode: ...
-
-@typing.type_check_only
-class _SetattrArgcountCodegenType(typing.Protocol):
-    def __call__(
-        self,
-        argcount: int,
-        slotted: bool,
-        /,
-        *,
-        funcname: str = ...
-    ) -> GeneratedCode: ...
-
-@typing.type_check_only
-class _BaseArgcountCodegenType(typing.Protocol):
-    def __call__(
-        self,
-        argcount: int,
-        /,
-        *,
-        funcname: str = ...
-    ) -> GeneratedCode: ...
-
-type _ArgcountCodegenType = _InitArgcountCodegenType | _SetattrArgcountCodegenType | _BaseArgcountCodegenType
 
 
 @typing.type_check_only
@@ -110,199 +72,12 @@ class AnnotationGathererProtocol(typing.Protocol, typing.Generic[_FieldType]):
         cls_annotations: None | dict[str, typing.Any],
     ) -> tuple[dict[str, _FieldType], dict[str, typing.Any]]: ...
 
-class GeneratedCode:
-    __slots__: tuple[str, ...]
-    source_code: str
-    globs: dict[str, typing.Any]
-    annotations: dict[str, typing.Any]
 
-    def __init__(
-        self,
-        source_code: str,
-        globs: dict[str, typing.Any] | None = ...,
-        annotations: dict[str, typing.Any] | None = ...,
-    ) -> None: ...
-    def __repr__(self) -> str: ...
-    def generate(self) -> types.FunctionType: ...
-
-class MethodMaker:
-    funcname: str
-    code_generator: _CodegenType
-    cached_generator: _CachedFunctionBuilder
-    decorator: None | Callable[[types.FunctionType], types.FunctionType]
-    def __init__(
-        self,
-        funcname: str,
-        code_generator: _CodegenType,
-        cached_generator: None | _CachedFunctionBuilder = ...,
-        decorator: None | Callable[[types.FunctionType], types.FunctionType] = ...,
-    ) -> None: ...
-    def __repr__(self) -> str: ...
-    def __get__(self, instance, cls) -> types.FunctionType: ...
-
-class _SignatureMaker:
-    def __get__(self, instance, cls=None) -> typing_extensions.Never: ...
-
-signature_maker: _SignatureMaker
-
-# Args
-def get_empty_args(cls: type) -> tuple[tuple[()]]: ...
-def get_init_args(cls: type) -> tuple[tuple[str, ...], bool, bool] | None: ...
-def get_compare_args(cls: type) -> tuple[tuple[str, ...]]: ...
-def get_repr_args(cls: type) -> tuple[tuple[str, ...]]: ...
-def get_replace_args(cls: type) -> tuple[tuple[str, ...]]: ...
-def get_frozen_setattr_args(cls: type) -> tuple[tuple[()], bool]: ...
-
-# Globals
-def get_init_globals(cls: type) -> dict[str, typing.Any]: ...
-def get_frozen_setattr_globals(cls: type) -> dict[str, typing.Any]: ...
-
-# Parameters
-type _FunctionParameterType = tuple[
-    tuple[str],
-    int,
-    int,
-    tuple[typing.Any],
-    dict[str, typing.Any],
-    dict[str, typing.Any],
-]
-
-def get_init_parameters(cls: type) -> _FunctionParameterType: ...
-
-# field names
-def get_counter_field_names(argcount: int) -> list[str]: ...
-
-class _CacheStats:
-    __slots__: tuple[str, str, str]
-    hits: int
-    misses: int
-    skips: int
-    @property
-    def hit_percent(self) -> float: ...
-    def __init__(self) -> None: ...
-    def __repr__(self) -> str: ...
-
-class _SimpleCache:
-    __slots__: tuple[str, ...]
-    _func: Callable[..., types.FunctionType]
-    _seed: dict[tuple, types.FunctionType]
-    _stats: _CacheStats
-    def __init__(
-        self,
-        func: types.FunctionType,
-        *,
-        cache_seed: dict[tuple, types.FunctionType] | None = ...,
-    ) -> None: ...
-    def __repr__(self) -> str: ...
-    def clear(
-        self,
-        new_cache: dict[tuple, types.FunctionType] | None = ...,
-    ) -> None: ...
-    @property
-    def stats(self) -> _CacheStats: ...
-    @property
-    def state(self) -> types.MappingProxyType[tuple, types.FunctionType]: ...
-    def __call__(self, *args, **kwargs) -> types.FunctionType: ...
-
-def _simple_cache(
-    *,
-    cache_seed: dict[tuple, types.FunctionType],
-) -> Callable[[types.FunctionType], _SimpleCache]: ...
-
-@typing.type_check_only
-class _CachedFunctionBuilder(typing.Protocol):
-    cache: _SimpleCache
-    def __call__(self, cls: type, funcname: str) -> types.FunctionType | None: ...
-
-def counter_to_class_generator(
-    counter_generator: _ArgcountCodegenType,
-    argument_getter: Callable[[type], tuple | None],
-    globals_getter: Callable[[type], dict[str, typing.Any]] | None = ...,
-    *,
-    cache: None | dict[tuple, types.FunctionType] = ...,
-    replace_strings: bool = ...,
-    param_updater: None | Callable[[type], _FunctionParameterType] = ...,
-) -> _CachedFunctionBuilder: ...
-
-# Actual generators
-# __init__
-def get_init_generator(
-    null: _NothingType = NOTHING, extra_code: None | list[str] = None
-) -> _CodegenType: ...
-def class_init_generator(cls: type, funcname: str = ...) -> GeneratedCode: ...
-def generic_init_generator(
-    field_names: list[str], frozen: bool, frozen_and_slotted: bool, *, funcname: str = ...
-) -> GeneratedCode: ...
-def _counter_init_generator(
-    argcount: int, frozen: bool, frozen_and_slotted: bool, /, *, funcname: str = ...
-) -> GeneratedCode: ...
-
-# __repr__
-def generic_repr_generator(
-    field_names: list[str], *, funcname: str = ...
-) -> GeneratedCode: ...
-def class_repr_generator(cls: type, funcname: str = ...) -> GeneratedCode: ...
-
-# __eq__
-def generic_eq_generator(
-    field_names: list[str], *, funcname: str = ...
-) -> GeneratedCode: ...
-def class_eq_generator(cls: type, funcname: str = ...) -> GeneratedCode: ...
-
-# __ge__, __gt__, __le__, __lt__
-def get_generic_order_generator(
-    field_names: list[str], operator: str, *, funcname: str
-) -> GeneratedCode: ...
-def get_class_order_generator(
-    cls: type, operator: str, *, funcname: str
-) -> GeneratedCode: ...
-def class_lt_generator(cls: type, funcname: str = ...) -> GeneratedCode: ...
-def class_le_generator(cls: type, funcname: str = ...) -> GeneratedCode: ...
-def class_gt_generator(cls: type, funcname: str = ...) -> GeneratedCode: ...
-def class_ge_generator(cls: type, funcname: str = ...) -> GeneratedCode: ...
-
-# __replace__
-def generic_replace_generator(
-    field_pairs: list[tuple[str, str]], *, funcname: str = ...
-) -> GeneratedCode: ...
-def class_replace_generator(cls: type, funcname: str = ...) -> GeneratedCode: ...
-
-# __setattr__ and __delattr__
-def generic_frozen_setattr_generator(
-    slotted: bool, *, funcname: str = ...
-) -> GeneratedCode: ...
-def class_frozen_setattr_generator(cls: type, funcname: str = ...) -> GeneratedCode: ...
-def generic_frozen_delattr_generator(*, funcname: str = ...) -> GeneratedCode: ...
-def class_frozen_delattr_generator(cls: type, funcname: str = ...) -> GeneratedCode: ...
-
-# __hash__
-def generic_hash_generator(
-    field_names: list[str], *, funcname: str = ...
-) -> GeneratedCode: ...
-def class_hash_generator(cls: type, funcname: str = ...) -> GeneratedCode: ...
-
-init_maker: MethodMaker
-repr_maker: MethodMaker
-eq_maker: MethodMaker
-lt_maker: MethodMaker
-le_maker: MethodMaker
-gt_maker: MethodMaker
-ge_maker: MethodMaker
-replace_maker: MethodMaker
-frozen_setattr_maker: MethodMaker
-frozen_delattr_maker: MethodMaker
-hash_maker: MethodMaker
 default_methods: frozenset[MethodMaker]
 
 _TypeT = typing.TypeVar("_TypeT", bound=type)
 
 # Construction functions
-def add_methods(
-    cls: type,
-    methods: Iterable[MethodMaker],
-    *,
-    internals: None | dict[str, typing.Any] = ...,
-) -> dict[str, MethodMaker]: ...
 @typing.overload
 def builder(
     cls: _TypeT,
@@ -361,7 +136,7 @@ class GatheredFields:
 class Field(metaclass=SlotMakerMeta):
     default: _NothingType | typing.Any
     default_factory: _NothingType | typing.Any
-    type: _NothingType | _py_type
+    _type: _NothingType | _private_type
     doc: None | str
     init: bool
     repr: bool
@@ -377,7 +152,7 @@ class Field(metaclass=SlotMakerMeta):
         *,
         default: _NothingType | typing.Any = ...,
         default_factory: _NothingType | typing.Any = ...,
-        type: _NothingType | _py_type = ...,
+        type: _NothingType | _private_type = ...,
         doc: None | str = ...,
         init: bool = ...,
         repr: bool = ...,
@@ -391,6 +166,9 @@ class Field(metaclass=SlotMakerMeta):
     def validate_field(self) -> None: ...
     @classmethod
     def from_field(cls, fld: Field, /, **kwargs: typing.Any) -> typing.Self: ...
+    @property
+    def type(self) -> _NothingType | _field_type: ...
+
 
 # These types only exist because type[Field] doesn't seem to resolve correctly
 # Technically they're wrong as `isinstance` gets used
