+++ src/ducktools/classbuilder/prefab.pyi
@@ -1,30 +1,51 @@
+# MIT License
+#
+# Copyright (c) 2024-2026 David C Ellis
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+__lazy_modules__: list[str]
+
 import typing
 from types import GenericAlias, MappingProxyType
 from typing_extensions import dataclass_transform
-
-__lazy_modules__: list[str]
 
 # Suppress weird pylance error
 from collections.abc import Callable  # type: ignore
 
 from . import (
-    NOTHING,
     Field,
     GathererProtocol,
-    GeneratedCode,
-    MethodMaker,
     SlotMakerMeta,
-    _SignatureMaker
 )
 
 from . import (
-    KW_ONLY as KW_ONLY,
     SlotFields as SlotFields,
     replace as replace,
 )
 
-# noinspection PyUnresolvedReferences
-from . import _NothingType
+from .constants import (
+    _NothingType,
+    NOTHING,
+    KW_ONLY as KW_ONLY,
+)
+from .methods import GeneratedCode, MethodMaker, _SignatureMaker
+
 
 _T = typing.TypeVar("_T")
